PAAC-668: Fixed unrelated messages to all input fields when user enter 5000000000 (Bigger value).

+++ app/form/AdjustedIncomeForm.scala
@@ -31,7 +31,7 @@ trait AdjustedIncomeForm extends AIFormFactory {
       Form[AdjustedIncomeModel](mapping(poundsDB(year))(toModel)(AdjustedIncomeModel.unapply))
   }
 
-  protected val toModel: (Int) => AdjustedIncomeModel = (i: Int) =>
+  protected val toModel: (Long) => AdjustedIncomeModel = (i: Long) =>
     AdjustedIncomeModel(BigDecimal(i))
 }
